fix: refactor frames app into frames.ts in examples
