+++ .changeset/thirty-dots-doubt.md
@@ -1,0 +1,5 @@
+---
+"framesjs-starter": patch
+---
+
+fix: refactor frames app into frames.ts in examples
