+++ examples/framesjs-starter/app/examples/new-api-mint-button/frames/frames.ts
@@ -1,0 +1,5 @@
+import { createFrames } from "frames.js/next";
+
+export const frames = createFrames({
+  basePath: "/examples/new-api-mint-button/frames",
+});
