+++ examples/framesjs-starter/app/examples/new-api-multi-protocol/frames/frames.ts
@@ -2,10 +2,10 @@
 import { farcasterHubContext, openframes } from "frames.js/middleware";
 import { createFrames } from "frames.js/next";
 import { getXmtpFrameMessage, isXmtpFrameActionPayload } from "frames.js/xmtp";
-import { DEFAULT_DEBUGGER_HUB_URL } from "../../debug";
+import { DEFAULT_DEBUGGER_HUB_URL } from "../../../debug";
 
 export const frames = createFrames({
-  basePath: "/examples/new-api-multi-protocol",
+  basePath: "/examples/new-api-multi-protocol/frames",
   initialState: {
     pageIndex: 0,
   },
