+++ examples/framesjs-starter/app/examples/new-api-multi-protocol/frames/route.tsx
@@ -1,5 +1,5 @@
 import { Button } from "frames.js/next";
-import { frames } from "../frames";
+import { frames } from "./frames";
 
 const handleRequest = frames(async (ctx) => {
   return {
